+++ CHANGES.md
@@ -1,5 +1,16 @@
 # Release notes
 
+## 1.1.0
+Jun 5, 2021
+
+- Created abstract **File** class
+- Created **TextFile** class
+- Added *\__str__* method for pretty representation of **Executor** class
+- Added *\__repr__* method for representation of **DatabaseManager** class
+- Added *\__repr__* method for representation of **FileManager** class
+- Added *\__repr__* method for representation of **LdapManager** class
+- Fix documentation for new abstract **File** class
+
 ## 1.0.0
 May 26, 2021